CDK optimizations (#14976)
